[MIG] port delivery_carrier_label_default to V10
