+++ delivery_carrier_label_default/__init__.py
@@ -1,1 +1,4 @@
+# -*- coding: utf-8 -*-
+# Copyright 2016 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from . import models
