--- delivery_carrier_label_default/__openerp__.py
@@ -1,20 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2013-2016 Camptocamp SA
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
-{'name': 'Default label for carrier labels',
- 'version': '9.0.1.0.0',
- 'author': "Camptocamp,Odoo Community Association (OCA)",
- 'maintainer': 'Camptocamp',
- 'category': 'Delivery',
- 'depends': [
-     'base_delivery_carrier_label',
- ],
- 'website': 'http://www.camptocamp.com/',
- 'data': [
-     'views/report_default_label.xml',
-     'views/reports.xml',
- ],
- 'installable': True,
- 'auto_install': False,
- 'license': 'AGPL-3',
- 'application': True}
