+++ delivery_carrier_label_default/models/__init__.py
@@ -1,1 +1,4 @@
-from . import stock
+# -*- coding: utf-8 -*-
+# Copyright 2016 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+from . import stock_picking
